services/pyllm/requirements.txt
# FastAPI: Python的现代化Web框架，用于构建API接口
fastapi

# Uvicorn: ASGI服务器，用于运行FastAPI应用
uvicorn

# Pydantic: 数据验证和设置管理库，用于定义API的请求和响应模型
pydantic

# OpenAI: OpenAI API客户端库，用于调用大语言模型API（兼容OpenAI和ModelScope等API）
openai

# python-dotenv: 用于加载.env文件中的环境变量
python-dotenv

langchain==1.1.3
langgraph==1.0.5
langgraph-prebuilt==1.0.5
langchain-nebius==0.1.3
rich==14.2.0
python-dotenv==1.2.1
langchain-community==0.4.1
langchain-tavily==0.2.14